# doc/api.yml
openapi: 3.0.1
info:
  version: '1.0.0'
  title: Conference API
  description: "This is an API to create and manage conferences.
               TODO: document endpoints for (1) implementing paper review and (2) operator/admin features"
security:
  - AuthToken: []
servers:
  - url: 'https://rnm.sg/conference/api/v0'
    description: Research and Methods conference API server'
paths:
...
  /conference/admin:
    get:
      summary: Get a list of all conferences
      operationId: getAllConference
      security:
        - AuthToken: [Operator]
      responses:
        200:
          description: A json object containing an array of conferences
          content:
            application/json:
              schema:
                type: object
                properties:
                  conferences:
                    type: array
                    items:
                      $ref: '#/components/schemas/Conference'
        401:
          $ref: '#/components/responses/401'
        403:
          $ref: '#/components/responses/403'
        422:
          $ref: '#/components/responses/422'
...
  /conference/{conferenceId}/join:
    post:
      summary: Join a conference
      description: 'Join the specified conference as the current user.
                    The membership type should be indicated in request body.'
      operationId: joinConference
      security:
        - AuthToken: []
      requestBody:
        required: true
        description: 'A JSON object that contains paper data.
                      A conference id of a conference open for submission is required.'
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/ConferenceId'
                - $ref: '#/components/schemas/Paper'
      responses:
        201:
          $ref: '#/components/responses/201'
          content:
            application/json:
              schema:
                type: object
                properties:
                  conferenceId:
                    type: string
                    format: uuid
                  title:
                    type: string
                  startDate:
                    type: string
                    format: date-time
                  lastDate:
                    type: string
                    format: date-time
        401:
          $ref: '#/components/responses/401'
        422:
          $ref: '#/components/responses/422'
...
components:
  schemas:
    Name:
      description: A person's name
      required:
        - firstName
        - lastName
      properties:
        firstName:
          type: string
        middleName:
          type: string
        lastName:
          type: string
    Address:
      type: object
      required:
        - addressLine1
        - city
        - zipCode
        - state
        - country
      properties:
        addressLine1:
          type: string
        addressLine2:
          type: string
        addressLine3:
          type: string
        city:
          type: string
        zipCode:
          type: string
        state:
          type: string
        country:
          type: string
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human readable description of the error.
...
    UserId:
      type: string
      format: uuid
...
    ConferenceId:
      type: object
      required:
        - cid
      properties:
        cid:
          type: string
          format: uuid
    Conference:
      type: object
      required:
        - name
        - startDate
        - lastDate
        - startPaperSubmissionDate
        - finalPaperSubmissionDate
        - paperSelectionDate
        - finalDraftSubmissionDate
        - venue
      properties:
        conferenceId:
          $ref: '#/components/schemas/ConferenceId'
        name:
          type: string
        startDate:
          type: string
          format: date-time
        lastDate:
          type: string
          format: date-time
        startPaperSubmissionDate:
          type: string
          format: date-time
        finalPaperSubmissionDate:
          type: string
          format: date-time
        paperSelectionDate:
          type: string
          format: date-time
        finalDraftSubmissionDate:
          type: string
          format: date-time
        venue:
          $ref: '#/components/schemas/Address'
        conferenceChair:
          type: string
        callForPapers:
          type: string
        hidden:
          type: boolean
    PaperId:
      type: string
      format: uuid
    Paper:
      type: object
      required:
        - title
        # - authors
        - presenter
        - uploader
      properties:
        paperId:
          $ref: '#/components/schemas/PaperId'
        title:
          type: string
        # authors:
        #   type: array
        #   items:
        #     $ref: '#/components/schemas/Author'
        presenterId:
          $ref: '#/components/schemas/UserId'
        uploaderId:
          $ref: '#/components/schemas/UserId'
        currentRevision:
          type: integer
...
    Author:
      type: object
      required:
        - name
        - email
      properties:
        name:
          $ref: '#/components/schemas/Name'
        email:
          type: string
          format: email
...
  securitySchemes:
    AuthToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-scopes:
        - User
        - ConferenceChair
        - Operator
  responses:
...
    201:
      description:
        'Requested resource was created. The content consists of a summarised form of the created resource.
        It includes a unique identifier to access the newly created resource.'
    401:
      description:
        'Missing authentication token. An authentication token is required to access the resource.
        A 404 maybe generated when accessing a protected resource without credentials if presence
        of the resource should not be known by unauthorized parties. For example, to prevent brute forcing paper ids.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    403:
      description:
        'Missing or invalid authentication token. An authentication token is required to access the resource.
        A 404 maybe generated when accessing a protected resource with invalid credentials if presence
        of the resource should not be known by unauthorized parties. For example, to prevent brute forcing paper ids.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    422:
      description:
        'One or more of the provided request parameters is invalid. Details can be found in the response body.
        In response to a POST request, this means that the requested resource was not created.
        In response to a PATCH request, this means that the requested resource was not altered.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'